+++ examples/04_model_loading/main.cpp
@@ -201,7 +201,7 @@ void on_resize(u32 width, u32 height)
     update_buffer(device, buffer_update);
 }
 
-void on_update(f64 delta_time)
+void on_update(f32 delta_time)
 {
     if (!command_buffers_recorded[ frame_index ])
     {
